+++ packages/wallets/exodus/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @swapkit/wallet-exodus
 
+## 1.2.36
+
+### Patch Changes
+
+- Updated dependencies [[`cc61490`](https://github.com/thorswap/SwapKit/commit/cc61490c52782468ab6e4494e5120df5b2f6d038), [`cc61490`](https://github.com/thorswap/SwapKit/commit/cc61490c52782468ab6e4494e5120df5b2f6d038), [`cc61490`](https://github.com/thorswap/SwapKit/commit/cc61490c52782468ab6e4494e5120df5b2f6d038)]:
+  - @swapkit/helpers@2.2.0
+  - @swapkit/toolbox-evm@1.7.10
+  - @swapkit/toolbox-utxo@1.2.10
+
 ## 1.2.35
 
 ### Patch Changes