+++ packages/wallets/polkadotjs/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @swapkit/wallet-polkadotjs
 
+## 1.0.52
+
+### Patch Changes
+
+- Updated dependencies [[`cc61490`](https://github.com/thorswap/SwapKit/commit/cc61490c52782468ab6e4494e5120df5b2f6d038), [`cc61490`](https://github.com/thorswap/SwapKit/commit/cc61490c52782468ab6e4494e5120df5b2f6d038), [`cc61490`](https://github.com/thorswap/SwapKit/commit/cc61490c52782468ab6e4494e5120df5b2f6d038)]:
+  - @swapkit/helpers@2.2.0
+  - @swapkit/toolbox-substrate@1.3.10
+
 ## 1.0.51
 
 ### Patch Changes